# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mausam 2.0 - Personalized Weather Intelligence",
  "description": "Mausam 2.0 - a Personalized Weather Intelligence Engine on official IMD/MoES data. Personalized homepage cards for the 'Mausam' app (SIH 2026 PS 26076).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register",
    "operationId": "register_api_v1_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/app__models__schemas__LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/logout-all": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Logout All",
    "description": "Revoke every token ever issued for the authenticated account.",
    "operationId": "logout_all_api_v1_auth_logout_all_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/account": {
   "delete": {
    "tags": [
     "auth"
    ],
    "summary": "Delete Account",
    "description": "Permanently erase the account and every associated record (DPDP erasure).\n\nRemoves the credential, the cloud profile backup, the in-app alert inbox\nand all registered push tokens. The client is expected to also wipe local\ndevice state after a successful response.",
    "operationId": "delete_account_api_v1_auth_account_delete",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Delete Account Api V1 Auth Account Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/request-otp": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Request Otp",
    "operationId": "request_otp_api_v1_auth_request_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OtpOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/verify-otp": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Verify Otp",
    "operationId": "verify_otp_api_v1_auth_verify_otp_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpVerifyRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/reset-password": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Reset Password",
    "description": "Recover an account with a reset OTP; invalidates all prior tokens.",
    "operationId": "reset_password_api_v1_auth_reset_password_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResetPasswordRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AuthOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "token": {
      "type": "string",
      "title": "Token"
     },
     "verified": {
      "type": "boolean",
      "title": "Verified",
      "default": true
     },
     "contact_verification_required": {
      "type": "boolean",
      "title": "Contact Verification Required",
      "default": false
     },
     "message": {
      "type": "string",
      "title": "Message",
      "default": "ok"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "token"
    ],
    "title": "AuthOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OtpOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "purpose": {
      "$ref": "#/components/schemas/OtpPurpose"
     },
     "sent_to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sent To"
     },
     "ttl_minutes": {
      "type": "integer",
      "title": "Ttl Minutes"
     },
     "dev_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dev Code"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "purpose",
     "ttl_minutes"
    ],
    "title": "OtpOut"
   },
   "OtpPurpose": {
    "type": "string",
    "enum": [
     "verify_email",
     "verify_phone",
     "reset_password"
    ],
    "title": "OtpPurpose"
   },
   "OtpRequest": {
    "properties": {
     "user_id": {
      "type": "string",
      "maxLength": 64,
      "minLength": 2,
      "pattern": "^[A-Za-z0-9_.-]+$",
      "title": "User Id"
     },
     "purpose": {
      "$ref": "#/components/schemas/OtpPurpose"
     },
     "contact": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 254
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "purpose"
    ],
    "title": "OtpRequest",
    "description": "Ask the server to generate + deliver a one-time-password.\n\n``contact`` is only needed when binding a contact during registration that\nwas not supplied there (it must match the account's own contact otherwise)."
   },
   "OtpVerifyRequest": {
    "properties": {
     "user_id": {
      "type": "string",
      "maxLength": 64,
      "minLength": 2,
      "pattern": "^[A-Za-z0-9_.-]+$",
      "title": "User Id"
     },
     "purpose": {
      "$ref": "#/components/schemas/OtpPurpose"
     },
     "code": {
      "type": "string",
      "maxLength": 8,
      "minLength": 4,
      "pattern": "^[0-9]{4,8}$",
      "title": "Code"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "purpose",
     "code"
    ],
    "title": "OtpVerifyRequest"
   },
...
   "RegisterRequest": {
    "properties": {
     "user_id": {
      "type": "string",
      "maxLength": 64,
      "minLength": 2,
      "pattern": "^[A-Za-z0-9_.-]+$",
      "title": "User Id"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 6,
      "title": "Password"
     },
     "email": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 254,
        "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 20,
        "pattern": "^\\+?[0-9]{10,15}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "password"
    ],
    "title": "RegisterRequest",
    "description": "Create a mobile account that owns a synced profile across devices.\n\n``email`` / ``phone`` are optional at registration but strongly advised:\nthey unlock OTP verification and self-serve password recovery (the only\nway to recover an account today)."
   },
   "ResetPasswordRequest": {
    "properties": {
     "user_id": {
      "type": "string",
      "maxLength": 64,
      "minLength": 2,
      "pattern": "^[A-Za-z0-9_.-]+$",
      "title": "User Id"
     },
     "code": {
      "type": "string",
      "maxLength": 8,
      "minLength": 4,
      "pattern": "^[0-9]{4,8}$",
      "title": "Code"
     },
     "new_password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 6,
      "title": "New Password"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "code",
     "new_password"
    ],
    "title": "ResetPasswordRequest",
    "description": "Recover an account using a reset OTP instead of the old password."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "app__models__schemas__LoginRequest": {
    "properties": {
     "user_id": {
      "type": "string",
      "maxLength": 64,
      "minLength": 2,
      "pattern": "^[A-Za-z0-9_.-]+$",
      "title": "User Id"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "password"
    ],
    "title": "LoginRequest"
   }
  }
 }
}